Refactor of ovn command into 2 subcomands
The `omc ovn subnets` command has been refactored into 2 subcommands:
- `omc ovn subnets` it contains only information about internal
  OVN subnets
- `omc ovn hostnetinfo` it contains only the information about the host
  network that OVN-Kubernetes needs to work (e.g. interface IPs, default
gateways...)

No functionality change other than the split, but this aims to be the
basis of future improvements.
